+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+Unreleased
+----------
+
+* Allow passing a custom ``transport`` to the underlying ``httpx`` client (e.g. a caching transport such as Hishel) - `PR 659`_
+
+.. _PR 659: https://github.com/martin-majlis/Wikipedia-API/pull/659
+
 0.15.0
 ----------